+++ server/test/unit/vitalsTest/checkHeartbeatTest.js
@@ -6,10 +6,10 @@ const sinonChai = require('sinon-chai')
 const rewire = require('rewire')
 
 // In-house dependencies
-const { helpers } = require('brave-alert-lib')
+const { factories, helpers } = require('brave-alert-lib')
 const db = require('../../../db/db')
 const redis = require('../../../db/redis')
-const { locationFactory } = require('../../../testingHelpers')
+const { locationFactory, sensorsVitalFactory } = require('../../../testingHelpers')
 
 const vitals = rewire('../../../vitals')
 
@@ -27,6 +27,10 @@ const excessiveRadarTimestamp = (currentRedisTimeInSeconds - radarThreshold - 1)
 const notExcessiveRadarTimestamp = (currentRedisTimeInSeconds - radarThreshold + 1) * 1000
 const excessiveDoorTimestamp = (currentRedisTimeInSeconds - doorThreshold - 1) * 1000
 const notExcessiveDoorTimestamp = (currentRedisTimeInSeconds - doorThreshold + 1) * 1000
+const excessiveRadarDate = new Date('2021-11-04T22:27:27.0248Z') // 61 seconds ago from currentDBDate
+const notExcessiveRadarDate = new Date('2021-11-04T22:27:29.0248Z') // 59 seconds ago from currentDBDate
+const excessiveDoorDate = new Date('2021-11-04T22:23:27.0248Z') // 301 seconds ago from currentDBDate
+const notExcessiveDoorDate = new Date('2021-11-04T22:23:29.0248Z') // 299 seconds ago from currentDBDate
 const resetReason = 'myResetReason'
 
 describe('vitals.js unit tests: checkHeartbeat', () => {
@@ -68,7 +72,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: notExcessiveDoorTimestamp })
 
-      sandbox.stub(redis, 'getLatestHeartbeat').throws('should not have called redis.getLatestHeartbeat')
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
 
       await vitals.checkHeartbeat()
     })
@@ -98,6 +102,87 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
     })
   })
 
+  describe('when an inactive device with a server-side state machine and no existing alerts notices that the latest radar message was longer than the threshold', () => {
+    beforeEach(async () => {
+      this.testLocation = locationFactory({ firmwareStateMachine: false, isActive: false, sentVitalsAlertAt: null })
+      sandbox.stub(db, 'getActiveServerStateMachineLocations').returns([this.testLocation])
+      sandbox.stub(db, 'getActiveFirmwareStateMachineLocations').returns([])
+
+      sandbox.stub(redis, 'getLatestXeThruSensorData').returns({ timestamp: excessiveRadarTimestamp })
+
+      sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: notExcessiveDoorTimestamp })
+
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
+
+      await vitals.checkHeartbeat()
+    })
+
+    it('should not send a radar disconnection message to Sentry', () => {
+      expect(helpers.logSentry).to.not.have.been.called
+    })
+
+    it('should not send an initial disconnection messages to the client', () => {
+      expect(this.sendDisconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not send any disconnection reminder messages to the client', () => {
+      expect(this.sendDisconnectionReminderStub).to.not.be.called
+    })
+
+    it('should not send any reconnection messages to the client', () => {
+      expect(this.sendReconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not update the sentVitalsAlertAt to the current time', () => {
+      expect(db.updateSentAlerts).to.not.be.called
+    })
+
+    it('should not log any errors', () => {
+      expect(helpers.logError).to.not.be.called
+    })
+  })
+
+  describe("when an inactive client's device with a server-side state machine and no existing alerts notices that the latest radar message was longer than the threshold", () => {
+    beforeEach(async () => {
+      const client = factories.clientFactory({ isActive: false })
+      this.testLocation = locationFactory({ firmwareStateMachine: false, isActive: true, sentVitalsAlertAt: null, client })
+      sandbox.stub(db, 'getActiveServerStateMachineLocations').returns([this.testLocation])
+      sandbox.stub(db, 'getActiveFirmwareStateMachineLocations').returns([])
+
+      sandbox.stub(redis, 'getLatestXeThruSensorData').returns({ timestamp: excessiveRadarTimestamp })
+
+      sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: notExcessiveDoorTimestamp })
+
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
+
+      await vitals.checkHeartbeat()
+    })
+
+    it('should not send a radar disconnection message to Sentry', () => {
+      expect(helpers.logSentry).to.not.have.been.called
+    })
+
+    it('should not send an initial disconnection messages to the client', () => {
+      expect(this.sendDisconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not send any disconnection reminder messages to the client', () => {
+      expect(this.sendDisconnectionReminderStub).to.not.be.called
+    })
+
+    it('should not send any reconnection messages to the client', () => {
+      expect(this.sendReconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not update the sentVitalsAlertAt to the current time', () => {
+      expect(db.updateSentAlerts).to.not.be.called
+    })
+
+    it('should not log any errors', () => {
+      expect(helpers.logError).to.not.be.called
+    })
+  })
+
   describe('when a device with a server-side state machine and no existing alerts notices that the latest door message was longer than the threshold', () => {
     beforeEach(async () => {
       this.testLocation = locationFactory({ firmwareStateMachine: false, isActive: true, sentVitalsAlertAt: null })
@@ -108,7 +193,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: excessiveDoorTimestamp })
 
-      sandbox.stub(redis, 'getLatestHeartbeat').throws('should not have called redis.getLatestHeartbeat')
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
 
       await vitals.checkHeartbeat()
     })
@@ -152,7 +237,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: notExcessiveDoorTimestamp })
 
-      sandbox.stub(redis, 'getLatestHeartbeat').throws('should not have called redis.getLatestHeartbeat')
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
 
       await vitals.checkHeartbeat()
     })
@@ -196,7 +281,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: notExcessiveDoorTimestamp })
 
-      sandbox.stub(redis, 'getLatestHeartbeat').throws('should not have called redis.getLatestHeartbeat')
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
 
       await vitals.checkHeartbeat()
     })
@@ -240,7 +325,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').returns({ timestamp: notExcessiveDoorTimestamp })
 
-      sandbox.stub(redis, 'getLatestHeartbeat').throws('should not have called redis.getLatestHeartbeat')
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').throws('should not have called db.getMostRecentSensorsVitalWithLocationid')
       await vitals.checkHeartbeat()
     })
 
@@ -279,13 +364,15 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
 
-      sandbox.stub(redis, 'getLatestHeartbeat').returns({ timestamp: excessiveRadarTimestamp, resetReason })
+      sandbox
+        .stub(db, 'getMostRecentSensorsVitalWithLocationid')
+        .returns(sensorsVitalFactory({ createdAt: excessiveRadarDate, doorLastSeenAt: notExcessiveDoorDate, resetReason }))
 
       await vitals.checkHeartbeat()
     })
 
     it('should send a system disconnection message to Sentry', () => {
-      expect(helpers.logSentry).to.have.been.calledOnceWithExactly(`System disconnected at ${this.testLocation.locationid}`)
+      expect(helpers.logSentry).to.have.been.calledOnceWithExactly(`Radar sensor down at ${this.testLocation.locationid}`)
     })
 
     it('should send an initial disconnection messages to the client', () => {
@@ -309,7 +396,134 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
     })
   })
 
-  describe('when a device with a firmware state machine and an existing alert is no longer exceeding radar thresholds', () => {
+  describe('when an inactive device with a firmware state machine and no existing alerts notices that the latest radar message was longer than the threshold', () => {
+    beforeEach(async () => {
+      this.testLocation = locationFactory({ firmwareStateMachine: true, isActive: false, sentVitalsAlertAt: null })
+      sandbox.stub(db, 'getActiveServerStateMachineLocations').returns([])
+      sandbox.stub(db, 'getActiveFirmwareStateMachineLocations').returns([this.testLocation])
+
+      sandbox.stub(redis, 'getLatestXeThruSensorData').throws('should not have called redis.getLatestXeThruSensorData')
+
+      sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
+
+      sandbox
+        .stub(db, 'getMostRecentSensorsVitalWithLocationid')
+        .returns(sensorsVitalFactory({ createdAt: excessiveRadarDate, doorLastSeenAt: notExcessiveDoorDate, resetReason }))
+
+      await vitals.checkHeartbeat()
+    })
+
+    it('should not send a system disconnection message to Sentry', () => {
+      expect(helpers.logSentry).to.not.have.been.called
+    })
+
+    it('should not send an initial disconnection messages to the client', () => {
+      expect(this.sendDisconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not send any disconnection reminder messages to the client', () => {
+      expect(this.sendDisconnectionReminderStub).to.not.be.called
+    })
+
+    it('should not send any reconnection messages to the client', () => {
+      expect(this.sendReconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not update the sentVitalsAlertAt to the current time', () => {
+      expect(db.updateSentAlerts).to.not.be.called
+    })
+
+    it('should not log any errors', () => {
+      expect(helpers.logError).to.not.be.called
+    })
+  })
+
+  describe("when an inactive client's device with a firmware state machine and no existing alerts notices that the latest radar message was longer than the threshold", () => {
+    beforeEach(async () => {
+      const client = factories.clientFactory({ isActive: false })
+      this.testLocation = locationFactory({ firmwareStateMachine: true, isActive: false, sentVitalsAlertAt: null, client })
+      sandbox.stub(db, 'getActiveServerStateMachineLocations').returns([])
+      sandbox.stub(db, 'getActiveFirmwareStateMachineLocations').returns([this.testLocation])
+
+      sandbox.stub(redis, 'getLatestXeThruSensorData').throws('should not have called redis.getLatestXeThruSensorData')
+
+      sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
+
+      sandbox
+        .stub(db, 'getMostRecentSensorsVitalWithLocationid')
+        .returns(sensorsVitalFactory({ createdAt: excessiveRadarDate, doorLastSeenAt: notExcessiveDoorDate, resetReason }))
+
+      await vitals.checkHeartbeat()
+    })
+
+    it('should not send a system disconnection message to Sentry', () => {
+      expect(helpers.logSentry).to.not.have.been.called
+    })
+
+    it('should not send an initial disconnection messages to the client', () => {
+      expect(this.sendDisconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not send any disconnection reminder messages to the client', () => {
+      expect(this.sendDisconnectionReminderStub).to.not.be.called
+    })
+
+    it('should not send any reconnection messages to the client', () => {
+      expect(this.sendReconnectionMessageStub).to.not.be.called
+    })
+
+    it('should not update the sentVitalsAlertAt to the current time', () => {
+      expect(db.updateSentAlerts).to.not.be.called
+    })
+
+    it('should not log any errors', () => {
+      expect(helpers.logError).to.not.be.called
+    })
+  })
+
+  describe('when a device with a firmware state machine and no existing alerts notices that the latest door message was longer than the threshold', () => {
+    beforeEach(async () => {
+      this.testLocation = locationFactory({ firmwareStateMachine: true, isActive: true, sentVitalsAlertAt: null })
+      sandbox.stub(db, 'getActiveServerStateMachineLocations').returns([])
+      sandbox.stub(db, 'getActiveFirmwareStateMachineLocations').returns([this.testLocation])
+
+      sandbox.stub(redis, 'getLatestXeThruSensorData').throws('should not have called redis.getLatestXeThruSensorData')
+
+      sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
+
+      sandbox
+        .stub(db, 'getMostRecentSensorsVitalWithLocationid')
+        .returns(sensorsVitalFactory({ createdAt: notExcessiveRadarDate, doorLastSeenAt: excessiveDoorDate, resetReason }))
+
+      await vitals.checkHeartbeat()
+    })
+
+    it('should send a system disconnection message to Sentry', () => {
+      expect(helpers.logSentry).to.have.been.calledOnceWithExactly(`Door sensor down at ${this.testLocation.locationid}`)
+    })
+
+    it('should send an initial disconnection messages to the client', () => {
+      expect(this.sendDisconnectionMessageStub).to.be.calledOnceWithExactly(this.testLocation.locationid, this.testLocation.displayName)
+    })
+
+    it('should not send any disconnection reminder messages to the client', () => {
+      expect(this.sendDisconnectionReminderStub).to.not.be.called
+    })
+
+    it('should not send any reconnection messages to the client', () => {
+      expect(this.sendReconnectionMessageStub).to.not.be.called
+    })
+
+    it('should update the sentVitalsAlertAt to the current time', () => {
+      expect(db.updateSentAlerts).to.be.calledOnceWithExactly(this.testLocation.locationid, true)
+    })
+
+    it('should not log any errors', () => {
+      expect(helpers.logError).to.not.be.called
+    })
+  })
+
+  describe('when a device with a firmware state machine and an existing alert is no longer exceeding the door or radar thresholds', () => {
     beforeEach(async () => {
       this.testLocation = locationFactory({
         firmwareStateMachine: true,
@@ -323,7 +537,9 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
 
-      sandbox.stub(redis, 'getLatestHeartbeat').returns({ timestamp: notExcessiveRadarTimestamp, resetReason })
+      sandbox
+        .stub(db, 'getMostRecentSensorsVitalWithLocationid')
+        .returns(sensorsVitalFactory({ createdAt: notExcessiveRadarDate, doorLastSeenAt: notExcessiveDoorDate, resetReason }))
 
       await vitals.checkHeartbeat()
     })
@@ -353,7 +569,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
     })
   })
 
-  describe('when a device with a firmware state machine and an existing alert is still exceeding the radar threshold but it has not yet exceeded the subsequent vitals threshold', () => {
+  describe('when a device with a firmware state machine and an existing alert is still exceeding the door or radar threshold but it has not yet exceeded the subsequent vitals threshold', () => {
     beforeEach(async () => {
       this.testLocation = locationFactory({
         firmwareStateMachine: true,
@@ -367,7 +583,9 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
 
-      sandbox.stub(redis, 'getLatestHeartbeat').returns({ timestamp: excessiveRadarTimestamp, resetReason })
+      sandbox
+        .stub(db, 'getMostRecentSensorsVitalWithLocationid')
+        .returns(sensorsVitalFactory({ createdAt: excessiveRadarDate, doorLastSeenAt: notExcessiveDoorDate, resetReason }))
 
       await vitals.checkHeartbeat()
     })
@@ -397,7 +615,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
     })
   })
 
-  describe('when a device with a firmware side state machine and an existing alert is still exceeding the radar threshold and has exceeded the subsequent vitals threshold', () => {
+  describe('when a device with a firmware side state machine and an existing alert is still exceeding the door or radar threshold and has exceeded the subsequent vitals threshold', () => {
     beforeEach(async () => {
       this.testLocation = locationFactory({
         firmwareStateMachine: true,
@@ -411,7 +629,7 @@ describe('vitals.js unit tests: checkHeartbeat', () => {
 
       sandbox.stub(redis, 'getLatestDoorSensorData').throws('should not have called redis.getLatestDoorSensorData')
 
-      sandbox.stub(redis, 'getLatestHeartbeat').returns({ timestamp: excessiveRadarTimestamp, resetReason })
+      sandbox.stub(db, 'getMostRecentSensorsVitalWithLocationid').returns({ createdAt: excessiveRadarDate, resetReason })
 
       await vitals.checkHeartbeat()
     })